@Viq111 2023-04-20
Benchmarks on
```
goos: darwin
goarch: arm64
pkg: github.com/DataDog/zstd
Apple M1 Max
```
with  [`mr`](https://github.com/DataDog/zstd/issues/1) payload:

```
                       │ test_before │             test_after             │
                       │   sec/op    │   sec/op     vs base               │
BulkCompress-10          969.1n ± 0%   992.5n ± 1%  +2.41% (p=0.000 n=20)
BulkDecompress-10        577.3n ± 0%   566.7n ± 0%  -1.84% (p=0.000 n=20)
CtxCompression-10        66.56m ± 0%   66.06m ± 0%  -0.75% (p=0.000 n=20)
CtxDecompression-10      9.444m ± 0%   9.300m ± 0%  -1.53% (p=0.000 n=20)
StreamCompression-10     73.45m ± 0%   72.59m ± 0%  -1.17% (p=0.000 n=20)
StreamDecompression-10   11.24m ± 0%   11.24m ± 1%       ~ (p=0.947 n=20)
Compression-10           66.59m ± 0%   66.03m ± 0%  -0.84% (p=0.000 n=20)
Decompression-10         9.458m ± 0%   9.303m ± 0%  -1.64% (p=0.000 n=20)
geomean                  1.918m        1.905m       -0.68%

                       │  test_before  │              test_after              │
                       │      B/s      │      B/s       vs base               │
BulkCompress-10           90.53Mi ± 0%    88.40Mi ± 1%  -2.36% (p=0.000 n=20)
BulkDecompress-10         94.17Mi ± 0%    95.92Mi ± 0%  +1.86% (p=0.000 n=20)
CtxCompression-10         142.9Mi ± 0%    143.9Mi ± 0%  +0.75% (p=0.000 n=20)
CtxDecompression-10      1006.8Mi ± 0%   1022.5Mi ± 0%  +1.56% (p=0.000 n=20)
StreamCompression-10      129.5Mi ± 0%    131.0Mi ± 0%  +1.18% (p=0.000 n=20)
StreamDecompression-10    845.9Mi ± 0%    846.0Mi ± 1%       ~ (p=0.952 n=20)
Compression-10            142.8Mi ± 0%    144.0Mi ± 0%  +0.85% (p=0.000 n=20)
Decompression-10         1005.3Mi ± 0%   1022.1Mi ± 0%  +1.67% (p=0.000 n=20)
geomean                   257.4Mi         259.2Mi       +0.68%
```

Overall there is no noticeable difference